Revert "Update QSR noc_non_blocking_api to custom instructions  (#38465)"
This reverts commit d7d68364a12a79d5e8be4c91794c301bff93059d.
