+++ tt_metal/hw/inc/internal/tt-2xx/quasar/stream_interface.h
@@ -7,9 +7,7 @@
 
 #include "noc_overlay_parameters.h"
 #include "noc_parameters.h"
-#if !defined(COMPILE_FOR_TRISC)
 #include "noc_nonblocking_api.h"
-#endif
 
 // Low-level chip-dependent stream/NOC functions
